build: fix IM compilation on bare-metal platforms
vexpress-a9 and zynq-a9 lack HTTPS/WebSocket infrastructure required
by Feishu and Telegram services.  Disable both IM integrations in
their cross.ini files instead of adding fragile #if guards.

telegram.c: replace PRId64 with %lld cast for portability across
toolchains where <inttypes.h> format macros may be unavailable.

feishu.c: revert unnecessary platform guards now that bare-metal
platforms disable the feature at the Meson level.

+++ claw/services/im/telegram.c
@@ -34,7 +34,6 @@
 
 #ifdef CONFIG_RTCLAW_TELEGRAM_ENABLE
 
-#include <inttypes.h>
 #include "osal/claw_net.h"
 #include "cJSON.h"
 
@@ -448,8 +447,8 @@ static void tg_poll_thread(void *arg)
     while (!claw_thread_should_exit()) {
         char body[128];
         snprintf(body, sizeof(body),
-                 "{\"offset\":%" PRId64 ",\"timeout\":%d}",
-                 ctx->update_offset, POLL_TIMEOUT_SEC);
+                 "{\"offset\":%lld,\"timeout\":%d}",
+                 (long long)ctx->update_offset, POLL_TIMEOUT_SEC);
 
         resp[0] = '\0';
         int ret = tg_api_post(ctx, "getUpdates", body, resp,
